Added a start to the css and clean up of the other files
I am removing old/undated comments from my html and js files. And we are rewording some of them. and to last we go truh to check if we have commented on what wer need to comment on and gotten our links down on paper for the help we used